Move Transformer modules to model directory
