+++ src/fairseq2/models/conformer/_block.py
@@ -12,9 +12,7 @@ from torch import Tensor
 from torch.nn import Dropout
 from typing_extensions import override
 
-from fairseq2.nn import LayerNorm
-from fairseq2.nn.padding import PaddingMask
-from fairseq2.nn.transformer import (
+from fairseq2.models.transformer import (
     AttentionMask,
     FeedForwardNetwork,
     LayerNormFactory,
@@ -22,6 +20,8 @@ from fairseq2.nn.transformer import (
     TransformerEncoderLayer,
     create_standard_layer_norm,
 )
+from fairseq2.nn import LayerNorm
+from fairseq2.nn.padding import PaddingMask
 from fairseq2.typing import DataType, Device
 
 # isort: split
