+++ src/fairseq2/models/llama/_factory.py
@@ -14,8 +14,18 @@ import torch.nn as nn
 from torch import Tensor
 
 from fairseq2.models.transformer import (
+    FeedForwardNetwork,
+    GLUFeedForwardNetwork,
+    MultiheadAttention,
+    StandardMultiheadAttention,
+    StandardTransformerDecoder,
+    StandardTransformerDecoderLayer,
+    TransformerDecoder,
+    TransformerDecoderLayer,
     TransformerEmbeddingFrontend,
     TransformerFrontend,
+    TransformerNormOrder,
+    create_default_sdpa,
 )
 from fairseq2.models.transformer_lm import TransformerLanguageModel
 from fairseq2.nn import (
@@ -28,18 +38,6 @@ from fairseq2.nn import (
     RotaryEncoder,
     StandardEmbedding,
     TiedProjection,
-)
-from fairseq2.nn.transformer import (
-    FeedForwardNetwork,
-    GLUFeedForwardNetwork,
-    MultiheadAttention,
-    StandardMultiheadAttention,
-    StandardTransformerDecoder,
-    StandardTransformerDecoderLayer,
-    TransformerDecoder,
-    TransformerDecoderLayer,
-    TransformerNormOrder,
-    create_default_sdpa,
 )
 from fairseq2.typing import DataType, Device
