+++ src/fairseq2/models/transformer/_attention_mask.py
