+++ src/fairseq2/models/transformer/_decoder.py
@@ -24,17 +24,17 @@ from fairseq2.typing import CPU, DataType, Device
 
 # isort: split
 
-from fairseq2.nn.transformer._attention_mask import (
+from fairseq2.models.transformer._attention_mask import (
     AttentionMaskFactory,
     CausalAttentionMaskFactory,
 )
-from fairseq2.nn.transformer._decoder_layer import TransformerDecoderLayer
-from fairseq2.nn.transformer._encoder import _record_drop_for_backward
-from fairseq2.nn.transformer._layer_norm import (
+from fairseq2.models.transformer._decoder_layer import TransformerDecoderLayer
+from fairseq2.models.transformer._encoder import _record_drop_for_backward
+from fairseq2.models.transformer._norm_order import TransformerNormOrder
+from fairseq2.models.transformer._normalization import (
     LayerNormFactory,
     create_standard_layer_norm,
 )
-from fairseq2.nn.transformer._norm_order import TransformerNormOrder
 
 
 class TransformerDecoder(LayerStack, ABC):
