+++ src/fairseq2/models/transformer/_ffn.py
