+++ src/fairseq2/models/transformer/_relative_attention.py
@@ -22,8 +22,11 @@ from fairseq2.typing import DataType, Device
 
 # isort: split
 
-from fairseq2.nn.transformer._attention import SDPA, create_default_sdpa
-from fairseq2.nn.transformer._attention_mask import AttentionMask, CustomAttentionMask
+from fairseq2.models.transformer._attention import SDPA, create_default_sdpa
+from fairseq2.models.transformer._attention_mask import (
+    AttentionMask,
+    CustomAttentionMask,
+)
 
 
 @final
