+++ src/fairseq2/models/wav2vec2/_config.py
@@ -10,7 +10,7 @@ from dataclasses import dataclass, field
 from typing import Final
 
 from fairseq2.context import RuntimeContext
-from fairseq2.nn.transformer import TransformerNormOrder
+from fairseq2.models.transformer import TransformerNormOrder
 from fairseq2.utils.validation import ValidationError, ValidationResult
 
 WAV2VEC2_MODEL_FAMILY: Final = "wav2vec2"
