+++ src/fairseq2/models/wav2vec2/_model.py
@@ -16,10 +16,10 @@ from torch.nn.functional import cross_entropy
 
 from fairseq2.error import InternalError
 from fairseq2.models.sequence import SequenceBatch
+from fairseq2.models.transformer import TransformerEncoder
 from fairseq2.nn import Linear
 from fairseq2.nn.ops import repeat_interleave
 from fairseq2.nn.padding import PaddingMask
-from fairseq2.nn.transformer import TransformerEncoder
 from fairseq2.typing import DataType, Device
 
 # isort: split
