+++ src/fairseq2/models/wav2vec2/asr/_checkpoint.py
@@ -11,8 +11,8 @@ from typing import cast
 from torch import Tensor
 from torch.nn.modules.utils import consume_prefix_in_state_dict_if_present
 
+from fairseq2.models.transformer import TransformerNormOrder
 from fairseq2.models.utils.checkpoint import convert_fairseq_checkpoint
-from fairseq2.nn.transformer import TransformerNormOrder
 
 # isort: split
