+++ src/fairseq2/models/wav2vec2/asr/_factory.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 
 import torch.nn as nn
 
+from fairseq2.models.transformer import TransformerEncoder
 from fairseq2.models.wav2vec2 import (
     StandardWav2Vec2Masker,
     Wav2Vec2EncoderFactory,
@@ -15,7 +16,6 @@ from fairseq2.models.wav2vec2 import (
     Wav2Vec2Masker,
 )
 from fairseq2.nn import Linear, Projection
-from fairseq2.nn.transformer import TransformerEncoder
 
 # isort: split
