+++ src/fairseq2/models/wav2vec2/asr/_model.py
@@ -13,9 +13,9 @@ from typing_extensions import override
 
 from fairseq2.models.asr import AsrModel, AsrModelOutput
 from fairseq2.models.sequence import SequenceBatch
+from fairseq2.models.transformer import TransformerEncoder
 from fairseq2.models.wav2vec2 import Wav2Vec2Frontend, Wav2Vec2Masker
 from fairseq2.nn import Projection
-from fairseq2.nn.transformer import TransformerEncoder
 from fairseq2.typing import DataType, Device
